Wire bottom navigation scaffold
