+++ chagok_flutter/lib/main.dart
@@ -1,10 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:chagok_flutter/screens/create_moment_screen.dart';
 import 'package:chagok_flutter/screens/photo_orientation_screen.dart';
-import 'package:chagok_flutter/screens/present_home_screen.dart';
-import 'package:chagok_flutter/screens/past_screen.dart';
-import 'package:chagok_flutter/screens/future_screen.dart';
 import 'package:chagok_flutter/theme/app_theme.dart';
+import 'package:chagok_flutter/widgets/main_navigation.dart';
 
 void main() {
   runApp(const ChagokApp());
@@ -18,7 +16,7 @@ class ChagokApp extends StatelessWidget {
     return MaterialApp(
       title: 'Chagok',
       theme: buildAppTheme(),
-      home: const HomeScaffold(),
+      home: const MainNavigation(),
       routes: {
         CreateMomentScreen.routeName: (_) => const CreateMomentScreen(),
       },
@@ -31,60 +29,6 @@ class ChagokApp extends StatelessWidget {
         }
         return null;
       },
-    );
-  }
-}
-
-class HomeScaffold extends StatefulWidget {
-  const HomeScaffold({super.key});
-
-  @override
-  State<HomeScaffold> createState() => _HomeScaffoldState();
-}
-
-class _HomeScaffoldState extends State<HomeScaffold> {
-  int _currentIndex = 1;
-
-  final List<Widget> _tabs = const [
-    PastScreen(),
-    PresentHomeScreen(),
-    FutureScreen(),
-  ];
-
-  void _onTabSelected(int index) {
-    setState(() {
-      _currentIndex = index;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: IndexedStack(
-        index: _currentIndex,
-        children: _tabs,
-      ),
-      bottomNavigationBar: BottomNavigationBar(
-        currentIndex: _currentIndex,
-        onTap: _onTabSelected,
-        selectedItemColor: AppColors.main,
-        unselectedItemColor: AppColors.textSecondary,
-        showUnselectedLabels: true,
-        items: const [
-          BottomNavigationBarItem(
-            icon: Icon(Icons.history_edu_outlined),
-            label: '과거',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.auto_awesome_outlined),
-            label: '현재',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.calendar_today_outlined),
-            label: '미래',
-          ),
-        ],
-      ),
     );
   }
 }
